app/requirements.txt
fastapi-cache2==0.2.2
fastapi==0.116.1
google-cloud-error-reporting==1.12.0
opentelemetry-api==1.36.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-grpc==0.57b0
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation==0.57b0
opentelemetry-propagator-gcp==1.9.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
python-multipart==0.0.20
uvicorn==0.35.0
pypdf2==3.0.1
google-cloud-aiplatform==1.109.0
langchain-google-vertexai==2.0.28
langchain==0.3.27
google-cloud-firestore==2.21.0
dotenv==0.9.9
langchain-community==0.3.27
langchain-openai==0.3.30
faiss-cpu==1.12.0
